pyproject.toml
[build-system]
requires = [ "poetry-core>=1.0.0",]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "optimus"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
repository = "https://github.com/valory-xyz/optimus"
keywords = [ "autonomy", "autonomous", "autonolas", "service", "multi", "agent",]
classifiers = [ "Environment :: Console", "Environment :: Web Environment", "Development Status :: 2 - Pre-Alpha", "Intended Audience :: Developers", "License :: OSI Approved :: Apache Software License", "Natural Language :: English", "Operating System :: MacOS", "Operating System :: Microsoft", "Operating System :: Unix", "Programming Language :: Python :: 3.8", "Programming Language :: Python :: 3.9", "Programming Language :: Python :: 3.10", "Programming Language :: Python :: 3.11", "Topic :: Communications", "Topic :: Internet", "Topic :: Scientific/Engineering", "Topic :: Software Development", "Topic :: System",]
[[tool.poetry.packages]]
include = "packages"

[tool.poetry.dependencies]
python = "<3.11.9,>=3.10"
open-autonomy = "==0.16.1"
open-aea-test-autonomy = "==0.16.1"
open-aea = "==1.57.0"
open-aea-ledger-ethereum = "==1.57.0"
typing_extensions = ">=3.10.0.2"
toml = "==0.10.2"
tomte = {version = "==0.2.17", extras = ["cli", "tests"]}
black = "==24.2.0"
isort = "==5.13.2"
grpcio = "==1.53.0"
asn1crypto = "<1.5.0,>=1.4.0"
py-ecc = "==6.0.0"
pytz = "==2022.2.1"
openapi-core = "==0.15.0"
openapi-spec-validator = "<0.5.0,>=0.4.0"
hypothesis = "==6.21.6"
requests = "<2.31.2,>=2.28.1"
py-multibase = "==1.0.3"
py-multicodec = "==0.2.1"
py-eth-sig-utils = "*"
open-aea-ledger-cosmos = "==1.57.0"
open-aea-ledger-solana = "==1.57.0"
protobuf = "<4.25.0,>=4.21.6"
web3 = "<7,>=6.0.0"
ipfshttpclient = "==0.8.0a2"
open-aea-cli-ipfs = "==1.57.0"
aiohttp = "<4.0.0,>=3.8.5"
certifi = "*"
multidict = "*"
ecdsa = ">=0.15"
eth_typing = "*"
hexbytes = "==0.3.1"
packaging = "*"
websocket_client = "<1,>=0.32.0"
pytest-asyncio = "*"
eth-utils = "==2.2.0"
eth-abi = "==4.0.0"
pycryptodome = "==3.18.0"
pytest = "==7.2.1"
urllib3 = "==1.26.16"
jsonschema = "<4.4.0,>=4.3.0"
pandas = ">=1.3.0"
pyalgotrade = "==0.20"
lyra-v2-client = ">=0.2.9"
balpy = {path = "third_party/balpy"}
numpy = "==1.26.1"
dateparser = ">=1.1.1"
multicaller = {path = "third_party/multicaller"}

[tool.poetry.group.dev.dependencies]
pytest-xprocess = ">=0.18.1,<0.19.0"
[tool.poetry.group.dev.dependencies.tomte]
version = ">=0.2.16"
extras = [ "cli", "tests",]
